fix: fix mistakes
