+++ 02/Evgeny Vasiukevich/src/App.tsx
@@ -1,26 +1,10 @@
-import React, { ChangeEvent, useState } from 'react';
-import HotelsBlock from './components/HomesGuestsLoves/HotelsBlock';
-import SearchForm from './components/UI/SearchForm/SearchForm';
-import { hotelsData } from './data/hotelsData';
+import React from 'react';
+import Hotels from './components/HomesGuestsLoves/Hotels';
 
 function App() {
-  const [filterHotels, setFilterHotels] = useState(hotelsData);
-  const [renderHotels, setRenderHotels] = useState(filterHotels);
-
-  const handleFilter = (event: ChangeEvent<HTMLInputElement>): void => {
-    setFilterHotels(hotelsData.filter(hotel => 
-      (hotel.name + hotel.city + hotel.country).toUpperCase().includes(event.target.value.toUpperCase())
-    ));
-  }
-
-  const handleRender = (): void => setRenderHotels(filterHotels);
-
   return (
     <div className="App">
-      <HotelsBlock hotels={hotelsData} title='Homes guests loves'>{}</HotelsBlock>
-      <HotelsBlock hotels={renderHotels} title='Available hotels'>
-        <SearchForm onFilter={handleFilter} onRender={handleRender}/>
-      </HotelsBlock>
+      <Hotels/>
     </div>
   );
 }
